+++ src/client/src/commons/menu/editor/mainSideNav.tsx
@@ -8,6 +8,7 @@ import AddIcon from "../../../../public/icons/add.svg?react";
 import UploadIcon from "../../../../public/icons/upload.svg?react";
 import SettingsIcon from "../../../../public/icons/settings.svg?react";
 import HelpIcon from "../../../../public/icons/help.svg?react";
+import LayersIcon from "../../../../public/icons/layers.svg?react";
 import { theme } from "../../../AppTheme";
 import { styled } from "@mui/system";
 import { ProfilePopup } from "../profilePopup";
@@ -75,6 +76,11 @@ const MainSideNav = ({
     setSideDrawerContent(DrawerContentTypes.NewBorehole);
   };
 
+  const handleToggleLayers = () => {
+    handleDrawer(DrawerContentTypes.CustomLayers);
+    setSideDrawerContent(DrawerContentTypes.CustomLayers);
+  };
+
   const handleDrawer = (buttonName: DrawerContentTypes) => {
     if (sideDrawerContent === buttonName) {
       toggleDrawer(!drawerOpen);
@@ -85,6 +91,7 @@ const MainSideNav = ({
 
   const isFilterPanelVisible = drawerOpen && sideDrawerContent === DrawerContentTypes.Filters;
   const isAddPanelVisible = drawerOpen && sideDrawerContent === DrawerContentTypes.NewBorehole;
+  const isLayersPanelVisible = drawerOpen && sideDrawerContent === DrawerContentTypes.CustomLayers;
 
   return (
     <Stack
@@ -123,6 +130,12 @@ const MainSideNav = ({
           }}
           disabled={user.data.roles.indexOf("EDIT") === -1}>
           <UploadIcon />
+        </StyledIconButton>
+        <StyledIconButton
+          data-cy="layers-button"
+          onClick={handleToggleLayers}
+          sx={isLayersPanelVisible ? selectedButtonStyle : {}}>
+          <LayersIcon />
         </StyledIconButton>
       </Stack>
       <Stack